Move classes to their own file. no issues. game worked just fine
