+++ main.rb
@@ -3,4 +3,4 @@ Dir[File.join(File.dirname(__FILE__), "lib", "**.rb")].each do |file|
 end
 
 # Game.setup
-Game.start
+WindowGame.start
